src/requirements.txt
-i https://pypi.org/simple
aniso8601==7.0.0
asgiref==3.2.10; python_version >= '3.5'
django-graphql-jwt==0.3.1
django==3.1
graphene-django==2.12.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
promise==2.3
psycopg2-binary==2.8.5
pyjwt==1.7.1
pytz==2020.1
rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
unidecode==1.1.1
